requirements.txt
supabase
psycopg2-binary
pytest
requests
python-dotenv
pytest-django
djangorestframework
pytest-reporter-html1
Pillow
pandas
numpy